Reduce the dependency on Activity from CordovaInterface
Most of the utility performed on the Activity instance can actually
be taken care of with a generic Context. In this case, the context
IS the Activity, but the same methods used are also available on
Context.

For instances of getActivity().runOnUiThread(), manually using
a Handler that was instantiated with the main Android looper.
This is effectively exactly what Activity#runOnUiThread does.

Added a convenience method on Cordova interface for posting to the
Handler.